build: warnings are errors, and six things that were hiding behind that
ASSISI_WARNINGS_AS_ERRORS existed, its own description said "recommended for
CI", and the base preset pinned it OFF — so every preset inherited a warning
set nothing enforced. Turning it on found six real things.

Four in our code:

  - App/tests/TestWorld.cpp: `baseline` is read only by a guard that tsan
    compiles out, so tsan built it unused. Marked maybe_unused where it is
    sampled, since it has to be read before the first world exists.
  - Core/src/BinaryCodec.cpp: a float widened to double implicitly while the
    64-bit cases beside it cast explicitly. The widening is intended; now it
    says so.
  - NetSync/MessageDispatch.hpp forward-declared ECS::Scene as `class` where
    it is defined as `struct`. Harmless on gcc/clang and a link error under
    the Microsoft ABI, where the class-key is part of the decorated name.
  - App/src/Application.cpp declared a local `Seconds` alias shadowing the
    file-scope one, with a comment claiming it was the only local alias. It
    had stopped being true.

One in the build system, which is the interesting one: App's test target
added TestBlueprintReplication.cpp through a $<BOOL:...> generator
expression. Assisi_apply_test_defaults reads SOURCES to attach per-source
warning opt-outs, an unevaluated genex comes back as its own literal text,
and so the clang opt-outs silently missed exactly that one file. Added with
if()/target_sources instead, so SOURCES holds filenames.

One suppressed rather than fixed, with the reason in the CMake: gcc's
-Warray-bounds mis-models a make_shared object whose derived part holds a
container and calls it "partly outside" an allocation of exactly that
object's size. Optimized builds only, test sources only, so a real
out-of-bounds in engine code is still an error.

Also finishes the ComponentMeta initializer lists in three test files, whose
"every member listed so -Wmissing-field-initializers stays quiet" comment had
gone stale when construct/getMutable were added.

Verified: gcc-debug, gcc-dev, gcc-ship, gcc-asan, gcc-tsan and clang-debug
all build with -Werror and pass 14/14.

+++ modules/Core/src/BinaryCodec.cpp
@@ -715,7 +715,10 @@ bool FieldsWithinBounds(std::span<const FieldMeta> fields, const void *object, s
         const void *address = FieldAddress(object, field.offset);
         switch (field.type)
         {
-        case FieldType::Float:  value = *static_cast<const float *>(address); break;
+        // Explicit, like the 64-bit cases below: the widening is intended — the
+        // whole comparison runs in double — and saying so is what keeps
+        // -Wdouble-promotion (clang) from reading it as an accident.
+        case FieldType::Float:  value = static_cast<double>(*static_cast<const float *>(address)); break;
         case FieldType::Double: value = *static_cast<const double *>(address); break;
         case FieldType::Int32:  value = *static_cast<const std::int32_t *>(address); break;
         case FieldType::UInt32: value = *static_cast<const std::uint32_t *>(address); break;
